+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Console output of `npm start` of generated applet shows correct URL of emulator
+
 ## [1.1.5] - 2023-01-02
 ### Fixed
 - Respect argument `--api-url` as priority over `SOS_API_URL` environment variable and default value